requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.6.0
Authlib==1.3.2
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
Django==3.2
exceptiongroup==1.2.2
fastapi==0.115.4
flake8==7.1.1
gitdb==4.0.11
GitPython==3.1.41
google==3.0.0
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Mako==1.3.6
MarkupSafe==3.0.2
mccabe==0.7.0
npx==0.1.6
numpy==2.1.3
packaging==24.1
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.29.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
pyparsing==3.2.0
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
pytz==2022.7.1
requests==2.32.3
rsa==4.9
setuptools==69.0.3
slack_sdk==3.33.4
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.4.3
starlette==0.41.2
tomli==2.0.2
tqdm==4.66.6
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
